extract function into util
