+++ app/core/SnapKeyring/utils/showDialog.ts
@@ -43,3 +43,34 @@ export async function showAccountNameSuggestionDialog(
     throw new Error(`Error occurred while showing name account dialog.\n${e}`);
   }
 }
+
+/**
+ * Show the account creation dialog for a given Snap.
+ * This function will start the approval flow, show the account creation dialog, and end the flow.
+ *
+ * @param snapId - Snap ID to show the account creation dialog for.
+ * @param messenger - The controller messenger instance.
+ * @returns The user's confirmation result.
+ */
+export async function showAccountCreationDialog(
+  snapId: string,
+  messenger: SnapKeyringBuilderMessenger,
+) {
+  try {
+    const confirmationResult = Boolean(
+      await messenger.call(
+        'ApprovalController:addRequest',
+        {
+          origin: snapId,
+          type: SNAP_MANAGE_ACCOUNTS_CONFIRMATION_TYPES.confirmAccountCreation,
+        },
+        true,
+      ),
+    );
+    return confirmationResult;
+  } catch (e) {
+    throw new Error(
+      `Error occurred while showing account creation dialog.\n${e}`,
+    );
+  }
+}
